% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceWMedians.R
\name{replaceNansWithMedian}
\alias{replaceNansWithMedian}
\title{Replaces NA values with median value of certain column}
\usage{
replaceNansWithMedian(df, colName)
}
\arguments{
\item{df}{The dataframe we wish to alter}

\item{colName}{Specific column name from dataframe we wish to alter}
}
\value{
Returns modified dataframe
}
\description{
Replaces NA values with median value of certain column
}
