% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCSV.R
\name{writeCSV}
\alias{writeCSV}
\title{Convert your data into a dataframe and write that dataframe to a .csv}
\usage{
writeCSV(testdata, predictstatement, filename = "yourAnswer")
}
\arguments{
\item{predictstatement}{Predictions generated from fit model}

\item{filename}{Desired new .csv file name}

\item{data}{Data to be put into a dataframe}
}
\value{
Nothing
}
\description{
Convert your data into a dataframe and write that dataframe to a .csv
}
